% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{coverage}
\alias{coverage}
\title{Empirical coverage of central prediction intervals}
\usage{
coverage(actual, lower, upper)
}
\arguments{
\item{actual}{Realised values.}

\item{lower, upper}{Interval bounds.}
}
\value{
Scalar coverage rate.
}
\description{
Empirical coverage of central prediction intervals
}
