% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{default_config}
\alias{default_config}
\title{Default experiment configuration}
\usage{
default_config()
}
\value{
A list with the same structure as \code{config/experiment.yml}.
}
\description{
Default experiment configuration
}
