% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_models}
\alias{get_models}
\title{Retrieve models by name}
\usage{
get_models(names = NULL)
}
\arguments{
\item{names}{Character vector of model names; \code{NULL} returns all.}
}
\value{
A named list of \code{stockcast_model} objects.
}
\description{
Retrieve models by name
}
